Inventory/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Flask-WTF==1.1.1
python-dotenv==1.0.0
openpyxl==3.1.2
pandas==2.0.3
email-validator==2.0.0.post2
